requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.0
streamlit>=1.31.0
python-dotenv>=1.0.0
openai>=1.12.0 